feat: added root posts page
Also moved each post to be behind `post/` and added description support
to posts.
