+++ .gitignore
@@ -4,9 +4,8 @@ Cargo.lock
 node_modules/
 pkg/
 
-# These are generated
+# This is generated
 static/tailwind.css
-.feed.xml
 
 # TODO Remove this in prod
 .blog/
