+++ Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 
 # Dependencies for the engine and the browser go here
 [dependencies]
-perseus = { version = "=0.4.0-beta.11", features = [ "hydrate" ] }
+perseus = { version = "=0.4.0-beta.11", features = [] }
 sycamore = "^0.8.1"
 serde = { version = "1", features = [ "derive" ] }
 serde_json = "1"
