+++ src/main.rs
@@ -13,6 +13,7 @@ pub fn main<G: Html>() -> PerseusApp<G> {
     PerseusApp::new()
         .template(crate::templates::index::get_template)
         .template(crate::templates::about::get_template)
+        .template(crate::templates::post::get_template)
         .template(crate::templates::posts::get_template)
         .template(crate::templates::tag::get_template)
         .template(crate::templates::series::get_template)
